Move useEffect logic inside the useAssistant hook
